pyproject.toml
[tool.poetry]
name = "bas-remote-python-v2"
# Poetry requires PEP 440-compliant versions for all projects.
version = "1.3"
license = "MIT"
description = "Python library, which allows you to automate Google Chrome browser via BAS software."
homepage = "https://github.com/CheshireCaat/bas-remote-python"
repository = "https://github.com/CheshireCaat/bas-remote-python"
documentation = "https://github.com/CheshireCaat/bas-remote-python/wiki"
authors = ["CheshireCaat <cheshirecat902@gmail.com>"]
readme = "README.md"
packages = [
    { include = "bas_remote" },
    { include = "tests" },
]
keywords = [
    "visual-programming-language",
    "browserautomationstudio",
    "bas-remote-control",
    "bas-remote-client",
    "bot-framework",
    "bas-remote",
    "bas-client",
    "automation",
    "chromium",
    "imacros",
    "windows",
    "desktop",
    "browser",
    "grabber",
    "poster",
    "macros",
    "bas",
    "ide",
    "cef",
    "bot",
]
classifiers = [
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/CheshireCaat/bas-remote-python/issues"

[tool.poetry.dependencies]
python = "^3.7"
dataclasses-json = "^0.5.7"
filelock = "^3.8.0"
websockets = "^10.3"
aiofiles = "^22.1.0"
aiohttp = "^3.8.1"
pyee = "^9.0.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-asyncio = "^0.19.0"
setuptools = "^65.3.0"
python-dotenv = "^0.21.0"
tox = "^3.25.1"
tox-poetry = "^0.4.1"
flake8 = "^5.0.4"
black = { git = "https://github.com/psf/black" }

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'

[tool.poetry.build]
script = "build.py"
generate-setup-file = false

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
